Actualización ubuntunning

+++ ubuntunning.sh
@@ -37,7 +37,7 @@ mkdir -p $TEMP_DIR
 cd $TEMP_DIR
 
 # Configurar sudoers para que no solicite contraseña
-echo -e "${YELLOW}[CONFIG] Configurando privilegios NOPASSWD para $REAL_USER...${NC}"
+echo -e "${BLUE}[CONFIG] Configurando privilegios NOPASSWD para $REAL_USER...${RESET}\n"
 sudo echo "$REAL_USER ALL=(ALL) NOPASSWD: ALL" | sudo tee /etc/sudoers.d/$REAL_USER > /dev/null
 sudo chmod 0440 /etc/sudoers.d/$REAL_USER
 
@@ -54,8 +54,8 @@ echo -e "\n${BLUE}[*] Instalando fuentes..${RESET}\n"
 mkdir -p $TEMP_DIR/fonts
 wget -q --show-progress https://github.com/ryanoasis/nerd-fonts/releases/download/v3.0.2/Hack.zip -O  $TEMP_DIR/Hack.zip
 wget -q --show-progress https://github.com/ryanoasis/nerd-fonts/releases/download/v3.0.2/JetBrainsMono.zip -O $TEMP_DIR/JetBrainsMono.zip
-unzip -q $TEMP_DIR/Hack.zip -d $TEMP_DIR/fonts
-unzip -q $TEMP_DIR/JetBrainsMono.zip -d $TEMP_DIR/fonts
+unzip -o -q $TEMP_DIR/Hack.zip -d $TEMP_DIR/fonts
+unzip -o -q $TEMP_DIR/JetBrainsMono.zip -d $TEMP_DIR/fonts
 mkdir -p ~/.local/share/fonts
 mv  $TEMP_DIR/fonts/*.ttf ~/.local/share/fonts/
 fc-cache -fv
@@ -81,43 +81,44 @@ cp -v $RPATH/config_files/color.ini ~/.config/kitty/
 cp -v $RPATH/config_files/kitty.conf ~/.config/kitty/
 
 # Instalación FortiClient VPN
-echo -e "${YELLOW}[DOWNLOAD] Descargando FortiClient VPN...${NC}"
+echo -e "\n${BLUE}[*] Descargando FortiClient VPN...${RESET}\n"
 wget -O forticlient_vpn.deb https://links.fortinet.com/forticlient/deb/vpnagent
-echo -e "${YELLOW}[INSTALL] Instalando FortiClient...${NC}"
+echo -e "\n${BLUE}[*] Instalando FortiClient...${RESET}\n"
 sudo apt install -y ./forticlient_vpn.deb
 
 # Instalación Chrome
-echo -e "${YELLOW}[INSTALL] Actualizando sistema e instalando dependencias base...${NC}"
+echo -e "\n${BLUE}[*] Actualizando sistema e instalando dependencias base...${RESET}\n"
 sudo apt install -y wget gnupg2 software-properties-common apt-transport-https ca-certificates curl
-echo -e "${YELLOW}[INSTALL] Configurando repositorio de Google Chrome...${NC}"
+echo -e "\n${BLUE}[*] Configurando repositorio de Google Chrome...${RESET}\n"
 wget -q -O - https://dl.google.com/linux/linux_signing_key.pub | sudo gpg --dearmor --yes -o /usr/share/keyrings/google-chrome.gpg
 sudo echo "deb [arch=amd64 signed-by=/usr/share/keyrings/google-chrome.gpg] http://dl.google.com/linux/chrome/deb/ stable main" | sudo tee /etc/apt/sources.list.d/google-chrome.list
 sudo apt update
 sudo apt install -y google-chrome-stable
 
 # Instalación maltego
-echo -e "${YELLOW}[DOWNLOAD] Descargando Maltego...${NC}"
+echo -e "\n${BLUE}[*] Descargando Maltego...${RESET}\n"
 # Nota: Verifica periódicamente si el enlace de versión sigue activo
 wget -O maltego.deb https://downloads.maltego.com/maltego-v4/linux/Maltego.v4.11.1.deb
 chmod +x maltego.deb
-echo -e "${YELLOW}[INSTALL] Instalando Maltego...${NC}"
+echo -e "\n${BLUE}[*] Instalando Maltego...${RESET}\n"
 sudo apt install -y ./maltego.deb
 
-#Instalación VSCode
-echo -e "${YELLOW}[DOWNLOAD] Visual Studio Code...${NC}"
-wget -O vscode.deb https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64
-chmod +x vscode.deb
-echo -e "${YELLOW}[INSTALL] Instalando Visual Studio Code...${NC}"
-sudo apt install -y ./vscode.deb
+# Instalación VSCode
+echo -e "\n${BLUE}[*] Visual Studio Code...${RESET}\n"
+wget -q --show-progress -O vscode.deb "https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64"
+if [ -f "vscode.deb" ]; then
+    echo -e "\n${BLUE}[*] Instalando Visual Studio Code...${RESET}\n"
+    sudo apt install -y ./vscode.deb
+else
+    echo -e "\n${RED}[ERROR] No se pudo descargar VS Code.${RESET}\n"
+fi
 
 #Instalacion MSF
-echo -e "${YELLOW}[DOWNLOAD] Descargando Metasploit Framework...${NC}"
-curl https://raw.githubusercontent.com/rapid7/metasploit-omnibus/master/config/templates/metasploit-framework-wrappers/msfupdate.erb > msfinstall && \
-  chmod 755 msfinstall && \
-  ./msfinstall
+echo -e "${BLUE}[DOWNLOAD] Descargando Metasploit Framework...${RESET}\n"
+curl https://raw.githubusercontent.com/rapid7/metasploit-omnibus/master/config/templates/metasploit-framework-wrappers/msfupdate.erb > msfinstall && chmod 755 msfinstall && ./msfinstall
 
 #Personalización del dock
-echo -e "${YELLOW}[CONFIG] Personalizando el Dock de GNOME para $REAL_USER...${NC}"
+echo -e "${BLUE}[CONFIG] Personalizando el Dock de GNOME para $REAL_USER...${NC}"
 ## Definir el bus para que root pueda hablar con la sesión de del usuario
 export DBUS_SESSION_BUS_ADDRESS="unix:path=/run/user/$(id -u "$REAL_USER")/bus"
 ## Definir el orden exacto de las aplicaciones
@@ -140,7 +141,7 @@ echo -e "${GREEN}------------------------------------------------${NC}"
 chsh -s /bin/zsh
 
 # Configurar Ctrl+Alt+T para abrir Kitty
-echo -e "${YELLOW}[CONFIG] Remapeando Ctrl+Alt+T a Kitty...${NC}"
+echo -e "${BLUE}[CONFIG] Remapeando Ctrl+Alt+T a Kitty...${NC}"
 gsettings set org.gnome.settings-daemon.plugins.media-keys terminal "['']"
 KEY_PATH="/org/gnome/settings-daemon/plugins/media-keys/custom-keybindings/custom0/"
 gsettings set org.gnome.settings-daemon.plugins.media-keys.custom-keybinding:$KEY_PATH name 'Abrir Kitty'
@@ -168,7 +169,7 @@ while true; do
             echo -e "${GREEN}[CLEANUP] Directorio eliminado.${NC}"
             break;;
         [Nn]* ) 
-            echo -e "${YELLOW}[INFO] Directorio conservado en $TEMP_DIR.${NC}"
+            echo -e "${BLUE}[INFO] Directorio conservado en $TEMP_DIR.${NC}"
             break;;
         * ) echo "Por favor responde s (sí) o n (no).";;
     esac
